add multi thread test and improve code logic
